requirements.txt
flask==3.0.3
line-bot-sdk==2.4.2
requests==2.31.0
schedule==1.2.2
pytz==2024.1
gunicorn==22.0.0
